NAS-118820: Moved AdminLayoutComponent
